fix(empty-state): guard orbit logic for empty tools and clean up dependencies
- remove unused useMemo import
- prevent NaN angles when tools.length === 0
- avoid recreating iconMap on each render by hoisting it to module scope
