+++ src/webview/EmptyState.tsx
@@ -1,116 +1,77 @@
 import React, { useMemo } from 'react';
 import { motion } from 'framer-motion';
 import { AlertTriangle } from 'lucide-react';
-import {
-  SiLangchain,
-  SiDvc,
-  SiGithub,
-  SiGitlab,
-  SiJenkins,
-  SiApacheairflow,
-  SiPrefect,
-  SiCircleci,
-  SiTravisci,
-  SiKedro,
-  SiUipath
-} from 'react-icons/si';
-import { IconType } from 'react-icons';
+import { getToolIcon } from './toolIcons';
 
 interface EmptyStateProps {
-  category: string;
-  tools: string[];
+    category: string;
+    tools: string[];
 }
 
-const getToolIcon = (toolName: string): IconType | null => {
-  const normalized = toolName.toLowerCase().replace(/[^a-z0-9]/g, '');
-
-  const iconMap: Record<string, IconType> = {
-    langchain: SiLangchain,
-    dvc: SiDvc,
-    githubaction: SiGithub,
-    gitlabci: SiGitlab,
-    jenkins: SiJenkins,
-    airflow: SiApacheairflow,
-    apacheairflow: SiApacheairflow,
-    prefect: SiPrefect,
-    circleci: SiCircleci,
-    travis: SiTravisci,
-    travisci: SiTravisci,
-    kedro: SiKedro,
-    uipath: SiUipath,
-  };
-
-  return iconMap[normalized] || null;
-};
-
 export const EmptyState: React.FC<EmptyStateProps> = ({ category, tools }) => {
-  // Use a fixed radius that works well with the design
-  const radius = 160;
+    // Use a fixed radius that works well with the design
+    const radius = 160;
 
-  return (
-    <div className="empty-state-container">
-      <AlertTriangle size={48} className="empty-state-icon" />
-      <h2 className="empty-state-title">
-        No pipeline files for category '{category}' were found.
-      </h2>
-      <p className="empty-state-subtitle">
-        We recommend you create a pipeline file for one of the following tools to enable visualization:
-      </p>
+    // Precompute planet configs to avoid redundant calculations
+    const planetConfigs = useMemo(
+        () =>
+            tools.length > 0
+                ? tools.map((tool, index) => ({
+                    tool,
+                    Icon: getToolIcon(tool),
+                    angle: (360 / tools.length) * index,
+                }))
+                : [],
+        [tools],
+    );
 
-      <div className="orbit-system-container">
-        {/* Central marker/sun */}
-        <div className="orbit-center-marker" />
+    return (
+        <div className="empty-state-container">
+            <AlertTriangle size={48} className="empty-state-icon" />
+            <h2 className="empty-state-title">
+                No pipeline files for category '{category}' were found.
+            </h2>
+            <p className="empty-state-subtitle">
+                We recommend you create a pipeline file for one of the following tools to enable visualization:
+            </p>
 
-        {/* The Orbit Path Visual */}
-        <div className="orbit-path-circle" />
+            <div className="orbit-system-container">
+                {/* Central marker/sun */}
+                <div className="orbit-center-marker" />
 
-        {/* Rotating Container */}
-        <motion.div
-          className="orbit-rotator"
-          animate={{ rotate: 360 }}
-          transition={{ duration: 40, repeat: Infinity, ease: "linear" }}
-        >
-          {tools.map((tool, index) => {
-            const Icon = getToolIcon(tool);
-            // Distribute tools evenly around the circle
-            const angle = (360 / tools.length) * index;
+                {/* The Orbit Path Visual */}
+                <div className="orbit-path-circle" />
 
-            return (
-              <motion.div
-                key={tool}
-                className="orbit-planet-slot"
-                style={{
-                  rotate: angle, // Static rotation to position the slot direction
-                }}
-              >
-                <div
-                  className="orbit-planet-arm"
-                  style={{
-                    width: radius, // Arm length pushes the planet out
-                  }}
-                >
-                  <motion.div
-                    className="orbit-planet-widget"
-                    // Counter-rotate to keep text upright as the parent rotates.
-                    // We animate from the initial negative angle to that angle minus 360 degrees
-                    // to complete a full counter-rotation cycle synchronized with the main rotator.
-                    animate={{ rotate: [-angle, -angle - 360] }}
-                    transition={{ duration: 40, repeat: Infinity, ease: "linear" }}
-                    style={{
-                      rotate: -angle
-                    }}
-                  >
-                    {Icon && <Icon className="planet-icon" />}
-                    <span className="planet-label">{tool}</span>
-                  </motion.div>
-                </div>
-              </motion.div>
-            );
-          })}
-        </motion.div>
-      </div>
+                {/* Rotating Container */}
+                {planetConfigs.length > 0 && (
+                    <motion.div
+                        className="orbit-rotator"
+                        animate={{ rotate: 360 }}
+                        transition={{ duration: 40, repeat: Infinity, ease: "linear" }}
+                    >
+                        {planetConfigs.map(({ tool, Icon, angle }) => (
+                            <div
+                                key={tool}
+                                className="orbit-planet"
+                                style={{
+                                    transform: `rotate(${angle}deg) translate(${radius}px)`,
+                                }}
+                            >
+                                <motion.div
+                                    className="orbit-planet-content"
+                                    animate={{ rotate: [-angle, -angle - 360] }}
+                                    transition={{ duration: 40, repeat: Infinity, ease: "linear" }}
+                                >
+                                    {Icon && <Icon className="planet-icon" />}
+                                    <span className="planet-label">{tool}</span>
+                                </motion.div>
+                            </div>
+                        ))}
+                    </motion.div>
+                )}
+            </div>
 
-      <style>{`
+            <style>{`
         .empty-state-container {
           width: 100%;
           flex: 1;
@@ -189,40 +150,16 @@ export const EmptyState: React.FC<EmptyStateProps> = ({ category, tools }) => {
             display: flex;
             align-items: center;
             justify-content: center;
-            /* The rotator spins the entire system */
         }
 
-        /* 
-           The Slot is centered in the container, rotated to point in the direction of the planet.
-        */
-        .orbit-planet-slot {
+        .orbit-planet {
             position: absolute;
             top: 50%;
             left: 50%;
-            width: 0;
-            height: 0;
-            display: flex;
-            align-items: center;
-            justify-content: flex-start;
-        }
-        
-        /* 
-           The Arm extends from the center (Slot) to the radius.
-           Flex content is aligned to flex-end so the Widget sits at the tip.
-        */
-        .orbit-planet-arm {
-            position: absolute;
-            left: 0;
-            top: 0;
-            height: 0;
-            transform-origin: left center;
-            display: flex;
-            align-items: center;
-            justify-content: flex-end;
-            pointer-events: none; /* Let clicks pass through the arm space */
+            transform-origin: center;
         }
 
-        .orbit-planet-widget {
+        .orbit-planet-content {
             background-color: #1f2332;
             border: 1px solid #3e3e42;
             padding: 10px 18px;
@@ -234,15 +171,11 @@ export const EmptyState: React.FC<EmptyStateProps> = ({ category, tools }) => {
             font-size: 13px;
             font-weight: 500;
             white-space: nowrap;
-            box-shadow: 0 4px 12px rgba(0,0,0,0.3);
-            pointer-events: auto; /* Re-enable pointer events for the widget */
-            
-            /* Center the widget on the end of the arm tip (radius point). */
-            /* Using translate(-50%, 0) centers it horizontally on the anchor point. */
-            transform: translate(-50%, 0); 
+            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+            transform: translate(-50%, -50%);
         }
         
-        .orbit-planet-widget:hover {
+        .orbit-planet-content:hover {
             border-color: #f20d63;
             box-shadow: 0 0 15px rgba(242, 13, 99, 0.4);
             color: #ffffff;
@@ -255,6 +188,6 @@ export const EmptyState: React.FC<EmptyStateProps> = ({ category, tools }) => {
             color: #f20d63;
         }
       `}</style>
-    </div>
-  );
+        </div>
+    );
 };
